PrincipalAxes now has only one template parameter
Now it internally uses double for accuracy considerations.
The type of floating point number can also be made a template parameter, but it is intentionally not designed like this, because large errors can be accumulated when dealing a large of number coordinates of different scales. So this class internally uses 'double'.